requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0

# NLP and Text Processing
nltk>=3.8.1
spacy>=3.5.0

# Computer Vision
opencv-python>=4.7.0
mediapipe>=0.10.0
Pillow>=9.0.0

# Backend
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Vector Database and RAG
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Local LLM
ollama>=0.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Training and Evaluation Dependencies
datasets>=2.14.0
accelerate>=0.20.0
evaluate>=0.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0 